+++ examples/my-app/CHANGELOG.md
@@ -1,5 +1,38 @@
 # my-app
 
+## 1.3.0
+
+### Minor Changes
+
+- fix eip1271 sign
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@0.19.0
+
+## 1.2.0
+
+### Minor Changes
+
+- CrossChain Subgraph Support
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/provider@0.18.0
+
+## 1.1.0
+
+### Minor Changes
+
+- 8568aa6: upgrade contract version
+
+### Patch Changes
+
+- Updated dependencies [8568aa6]
+  - @0xsodium/provider@0.17.0
+
 ## 1.0.2
 
 ### Patch Changes