+++ packages/wallet/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @0xsodium/wallet
 
+## 0.19.0
+
+### Minor Changes
+
+- fix eip1271 sign
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@0.19.0
+  - @0xsodium/config@0.19.0
+  - @0xsodium/guard@0.19.0
+  - @0xsodium/network@0.19.0
+  - @0xsodium/sdk4337@0.19.0
+  - @0xsodium/transactions@0.19.0
+  - @0xsodium/utils@0.19.0
+
+## 0.18.0
+
+### Minor Changes
+
+- CrossChain Subgraph Support
+
+### Patch Changes
+
+- Updated dependencies
+  - @0xsodium/abi@0.18.0
+  - @0xsodium/config@0.18.0
+  - @0xsodium/guard@0.18.0
+  - @0xsodium/network@0.18.0
+  - @0xsodium/sdk4337@0.18.0
+  - @0xsodium/transactions@0.18.0
+  - @0xsodium/utils@0.18.0
+
+## 0.17.0
+
+### Minor Changes
+
+- 8568aa6: upgrade contract version
+
+### Patch Changes
+
+- Updated dependencies [8568aa6]
+  - @0xsodium/config@0.17.0
+  - @0xsodium/abi@0.17.0
+  - @0xsodium/guard@0.17.0
+  - @0xsodium/network@0.17.0
+  - @0xsodium/sdk4337@0.17.0
+  - @0xsodium/transactions@0.17.0
+  - @0xsodium/utils@0.17.0
+
 ## 0.16.0
 
 ### Minor Changes